Remove original code, move next code
