--- magik-squid/src/main/java/nl/ramsolutions/sw/magik/formatting/NullIndentStrategy.java
@@ -1,30 +1,0 @@
-package nl.ramsolutions.sw.magik.formatting;
-
-import com.sonar.sslr.api.AstNode;
-import com.sonar.sslr.api.Token;
-
-/** Null indent strategy. */
-public class NullIndentStrategy extends IndentStrategy {
-
-  public static final String NAME = "null";
-
-  public NullIndentStrategy(final FormattingOptions options) {
-    super(options);
-  }
-
-  @Override
-  public void setTokenColumnTracker(final TokenColumnTracker tokenColumnTracker) {
-    // Null indent strategy does not use the token column tracker.
-  }
-
-  @Override
-  public String getStrategyName() {
-    return NullIndentStrategy.NAME;
-  }
-
-  @Override
-  public String indentFor(final Token token, final AstNode currentNode) {
-    final int indentSize = token.getColumn();
-    return this.indentString(indentSize);
-  }
-}
