--- magik-squid/src/main/java/nl/ramsolutions/sw/magik/formatting/next/package-info.java
@@ -1,4 +1,0 @@
-/** Defaults. */
-@edu.umd.cs.findbugs.annotations.DefaultAnnotationForParameters(
-    edu.umd.cs.findbugs.annotations.NonNull.class)
-package nl.ramsolutions.sw.magik.formatting.next;
